+++ Components/HierarchicalPathTest.php
@@ -12,7 +12,6 @@
 namespace League\Uri\Components;
 
 use ArrayIterator;
-use GuzzleHttp\Psr7\Utils;
 use League\Uri\Contracts\UriInterface;
 use League\Uri\Exceptions\OffsetOutOfBounds;
 use League\Uri\Exceptions\SyntaxError;
@@ -85,19 +84,6 @@ final class HierarchicalPathTest extends TestCase
             "Don't encode unreserved chars or sub-delimiters" => ["/$unreserved", "/$unreserved"],
             'Encoded unreserved chars are not decoded' => ['/p%61th', '/p%61th'],
         ];
-    }
-
-    /**
-     * @covers ::createFromPath
-     * @covers ::withContent
-     */
-    public function testWithContent(): void
-    {
-        $str = '/path/to/the/sky';
-        $path = HierarchicalPath::createFromPath(new Path($str));
-
-        self::assertSame($path, $path->withContent($str));
-        self::assertNotEquals($path, $path->withContent('foo/bar'));
     }
 
     /**
@@ -824,7 +810,7 @@ final class HierarchicalPathTest extends TestCase
      */
     public function testCreateFromUriWithPSR7Implementation(): void
     {
-        $uri = Utils::uriFor('http://example.com')
+        $uri = Uri::createFromString('http://example.com')
             ->withPath('/path');
 
         self::assertSame('/path', $uri->getPath());