+++ Components/IpAddressTest.php
@@ -23,20 +23,6 @@ use Stringable;
 final class IpAddressTest extends TestCase
 {
     /**
-     * @covers ::__construct
-     * @covers ::withContent
-     * @covers ::isValidIpv6Hostname
-     */
-    public function testWithContent(): void
-    {
-        $host = new Host('127.0.0.1');
-
-        self::assertSame($host, $host->withContent('127.0.0.1'));
-        self::assertSame($host, $host->withContent($host));
-        self::assertNotSame($host, $host->withContent('[::1]'));
-    }
-
-    /**
      * Test valid IpAddress.
      *
      * @dataProvider validIpAddressProvider