+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 All Notable changes to `League\Uri\Components` will be documented in this file
 
-## [2.5.0](https://github.com/thephpleague/uri-components/compare/2.4.1...2.5.0) - TBD
+## [Next](https://github.com/thephpleague/uri-components/compare/2.4.1...master) - TBD
 
 ### Added
 
-- `value` method to replace deprecated `getContent` method.
+- `value` method to replace the removed `getContent` method.
 - requires `League\Uri` package.
+- `UserInfo::createFromString` named constructor
 
 ### Fixed
 
@@ -16,11 +17,12 @@ All Notable changes to `League\Uri\Components` will be documented in this file
 
 ### Deprecated
 
-- `withContent` and `getContent` methods
+- None
 
 ### Removed
 
 - PHP7 and PHP8.0 support
+- `withContent` and `getContent` methods
 
 ## [2.4.1](https://github.com/thephpleague/uri-components/compare/2.4.0...2.4.1) - 2022-05-26